requirements.txt
deta==1.1.0
Pillow==9.5.0
Pillow==9.5.0
plotly==5.15.0
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.24.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
